chore: set typescript
