+++ package.json
@@ -36,9 +36,11 @@
     ]
   },
   "devDependencies": {
+    "@tsconfig/docusaurus": "^1.0.7",
     "@types/react": "^17.0.43",
     "copyfiles": "^2.4.1",
-    "gh-pages": "^5.0.0"
+    "gh-pages": "^5.0.0",
+    "typescript": "^5.1.3"
   },
   "license": "UNLICENSED"
 }
